Seperate out version comparison
